feat(lite/xapi-subscriptions): add an `immediate` option + TypeScript enhancement
- `subscribe({ immediate: false })` allows to defer the subscription
- Extracted typing to its own file
- Enhanced `subscribe` signature with overloading
- Enhanced Host/VM stores `subscribe` signature and typing
